BotFDS2.0/BotFDS2.0/BotFDS2.0/src/requirements.txt
# Зависимости
aiogram==3.0.0b7
aiohttp==3.8.5
aiosqlite>=0.18.0
python-dotenv>=0.19.0